Numeric.AD.Internal.Combinators

Description

Combinators used internally by Numeric.AD

Documentation

zipWithT :: (Foldable f, Traversable g) => (a -> b -> c) -> f a -> g b -> g c

Zip a Foldable f with a Traversable g assuming f has at least as many entries as g.

zipWithDefaultT :: (Foldable f, Traversable g) => a -> (a -> b -> c) -> f a -> g b -> g c

Zip a Foldable f with a Traversable g assuming f, using a default value after f is exhausted.

withPrimal :: Jacobian t => t -> Scalar t -> t

Used internally to define various Enum combinators.

fromBy :: Jacobian t => t -> t -> Int -> Scalar t -> t

Used internally to define various Enum combinators.