Numeric.AD.Jet

Description

Documentation

data Jet f a

A Jet is a tower of all (higher order) partial derivatives of a function

At each step, a Jet f is wrapped in another layer worth of f.

a :- f a :- f (f a) :- f (f (f a)) :- ...

Constructors

a :- (Jet f (f a)) infixl 3

headJet :: Jet f a -> a

Take the head of a Jet.

tailJet :: Jet f a -> Jet f (f a)

Take the tail of a Jet.

jet :: Functor f => Cofree f a -> Jet f a

Construct a Jet by unzipping the layers of a Cofree Comonad.