Numeric.AD.Mode

Description

AD modes

class (Num t, Num (Scalar t)) => Mode t where

Minimal complete definition

auto

Associated Types

type Scalar t

Methods

isKnownConstant :: t -> Bool

allowed to return False for items with a zero derivative, but we'll give more NaNs than strictly necessary

isKnownZero :: t -> Bool

allowed to return False for zero, but we give more NaN's than strictly necessary then

auto :: Scalar t -> t

Embed a constant

(*^) :: Scalar t -> t -> t infixr 7

Scalar-vector multiplication

(^*) :: t -> Scalar t -> t infixl 7

Vector-scalar multiplication

(^/) :: Fractional (Scalar t) => t -> Scalar t -> t infixr 7

Scalar division

zero :: t

zero = lift 0