Numeric.AD.Variadic

Description

Variadic combinators for variadic mixed-mode automatic differentiation.

Unfortunately, variadicity comes at the expense of being able to use quantification to avoid sensitivity confusion, so be careful when counting the number of lift you use when taking the gradient of a function that takes gradients!

Reverse-mode variadic gradient

class Num a => Grad i o o' a | i -> a o o', o -> a i o', o' -> a i o

Minimal complete definition

pack, unpack, unpack'

vgrad :: Grad i o o' a => i -> o

vgrad' :: Grad i o o' a => i -> o'

Sparse forward mode variadic jet

class Num a => Grads i o a | i -> a o, o -> a i

Minimal complete definition

packs, unpacks

vgrads :: Grads i o a => i -> o