Numeric.AD.Variadic.Sparse

Description

Variadic combinators for sparse forward mode automatic differentiation.

Unfortunately, variadicity comes at the expense of being able to use quantification to avoid sensitivity confusion, so be careful when counting the number of lift you use when taking the gradient of a function that takes gradients!

Unsafe Variadic Gradient

class Num a => Grad i o o' a | i -> a o o', o -> a i o', o' -> a i o

vgrad :: Grad i o o' a => i -> o

vgrad' :: Grad i o o' a => i -> o'

class Num a => Grads i o a | i -> a o, o -> a i

vgrads :: Grads i o a => i -> o