Numeric.AD.Internal.Classes

Description

AD modes

class Num (Scalar t) => Mode t where

Minimal complete definition

auto, (<+>)

Methods

isKnownConstant :: t -> Bool

allowed to return False for items with a zero derivative, but we'll give more NaNs than strictly necessary

isKnownZero :: t -> Bool

allowed to return False for zero, but we give more NaN's than strictly necessary then

auto :: Scalar t -> t

Embed a constant

(<+>) :: Num (Scalar t) => t -> t -> t infixr 6

Vector sum

(*^) :: Scalar t -> t -> t infixr 7

Scalar-vector multiplication

(^*) :: t -> Scalar t -> t infixl 7

Vector-scalar multiplication

(^/) :: (Num t, Fractional (Scalar t)) => t -> Scalar t -> t infixr 7

Scalar division

zero :: t

'zero' = 'lift' 0

one :: Mode t => t

negOne :: Mode t => t

Automatically Deriving AD

class (Mode t, Mode (D t), Num (D t)) => Jacobian t where

Jacobian is used by deriveMode but is not exposed via Mode to prevent its abuse by end users via the AD data type.

Associated Types

type D t :: *

Methods

unary :: (Scalar t -> Scalar t) -> D t -> t -> t

lift1 :: (Scalar t -> Scalar t) -> (D t -> D t) -> t -> t

lift1_ :: (Scalar t -> Scalar t) -> (D t -> D t -> D t) -> t -> t

binary :: (Scalar t -> Scalar t -> Scalar t) -> D t -> D t -> t -> t -> t

lift2 :: (Scalar t -> Scalar t -> Scalar t) -> (D t -> D t -> (D t, D t)) -> t -> t -> t

lift2_ :: (Scalar t -> Scalar t -> Scalar t) -> (D t -> D t -> D t -> (D t, D t)) -> t -> t -> t

type family Scalar t

withPrimal :: (Jacobian t, Scalar t ~ Scalar (D t)) => t -> Scalar t -> t

fromBy :: (Jacobian t, Scalar t ~ Scalar (D t)) => t -> t -> Int -> Scalar t -> t