Numeric.AD.Internal.Kahn

Description

This module provides reverse-mode Automatic Differentiation implementation using linear time topological sorting after the fact.

For this form of reverse-mode AD we use StableName to recover sharing information from the tape to avoid combinatorial explosion, and thus run asymptotically faster than it could without such sharing information, but the use of side-effects contained herein is benign.

Documentation

newtype Kahn a s

Kahn is a Mode using reverse-mode automatic differentiation that provides fast diffFU, diff2FU, grad, grad2 and a fast jacobian when you have a significantly smaller number of outputs than inputs.

Constructors

Kahn (Tape a (Kahn a s))

data Tape a t

A Tape records the information needed back propagate from the output to each input during reverse Mode AD.

Constructors

Zero

Lift !a

Var !a !Int

Binary !a a a t t

Unary !a a t

partials :: forall s a. Num a => Kahn a s -> [(Int, a)]

This returns a list of contributions to the partials. The variable ids returned in the list are likely not unique!

partialArray :: Num a => (Int, Int) -> Kahn a s -> Array Int a

Return an Array of partials given bounds for the variable IDs.

partialMap :: Num a => Kahn a s -> IntMap a

Return an IntMap of sparse partials

derivative :: Num a => Kahn a s -> a

derivative' :: Num a => Kahn a s -> (a, a)

vgrad :: Grad i o o' a => i -> o

vgrad' :: Grad i o o' a => i -> o'

class Num a => Grad i o o' a | i -> a o o', o -> a i o', o' -> a i o where

Methods

pack :: i -> [Kahn a ()] -> Kahn a ()

unpack :: ([a] -> [a]) -> o

unpack' :: ([a] -> (a, [a])) -> o'

bind :: Traversable f => f a -> (f (Kahn a s), (Int, Int))

unbind :: Functor f => f (Kahn a s) -> Array Int a -> f a

unbindMap :: (Functor f, Num a) => f (Kahn a s) -> IntMap a -> f a

unbindWith :: (Functor f, Num a) => (a -> b -> c) -> f (Kahn a s) -> Array Int b -> f c

unbindMapWithDefault :: (Functor f, Num a) => b -> (a -> b -> c) -> f (Kahn a s) -> IntMap b -> f c

primal :: Num a => Kahn a s -> a

var :: a -> Int -> Kahn a s

varId :: Kahn a s -> Int