Numeric.AD.Internal.Sparse

Description

Unsafe and often partial combinators intended for internal usage.

Handle with care.

Documentation

newtype Index

Constructors

Index (IntMap Int)

emptyIndex :: Index

addToIndex :: Int -> Index -> Index

indices :: Index -> [Int]

data Sparse a

We only store partials in sorted order, so the map contained in a partial will only contain partials with equal or greater keys to that of the map in which it was found. This should be key for efficiently computing sparse hessians. there are only (n + k - 1) choose k distinct nth partial derivatives of a function with k inputs.

Constructors

Sparse !a (IntMap (Sparse a))

Zero

apply :: (Traversable f, Num a) => (f (Sparse a) -> b) -> f a -> b

vars :: (Traversable f, Num a) => f a -> f (Sparse a)

d :: (Traversable f, Num a) => f b -> Sparse a -> f a

d' :: (Traversable f, Num a) => f a -> Sparse a -> (a, f a)

ds :: (Traversable f, Num a) => f b -> Sparse a -> Cofree f a

skeleton :: Traversable f => f a -> f Int

spartial :: Num a => [Int] -> Sparse a -> Maybe a

partial :: Num a => [Int] -> Sparse a -> a

vgrad :: Grad i o o' a => i -> o

vgrad' :: Grad i o o' a => i -> o'

vgrads :: Grads i o a => i -> o

class Num a => Grad i o o' a | i -> a o o', o -> a i o', o' -> a i o where

Methods

pack :: i -> [Sparse a] -> Sparse a

unpack :: ([a] -> [a]) -> o

unpack' :: ([a] -> (a, [a])) -> o'

class Num a => Grads i o a | i -> a o, o -> a i where

Methods

packs :: i -> [Sparse a] -> Sparse a

unpacks :: ([a] -> Cofree [] a) -> o