Numeric.AD.Internal.Sparse

Description

Unsafe and often partial combinators intended for internal usage.

Handle with care.

Documentation

newtype Monomial

Constructors

Monomial (IntMap Int)

emptyMonomial :: Monomial

addToMonomial :: Int -> Monomial -> Monomial

indices :: Monomial -> [Int]

data Sparse a

We only store partials in sorted order, so the map contained in a partial will only contain partials with equal or greater keys to that of the map in which it was found. This should be key for efficiently computing sparse hessians. there are only (n + k - 1) choose (k - 1) distinct nth partial derivatives of a function with k inputs.

Constructors

Sparse !a (IntMap (Sparse a))

Zero

apply :: (Traversable f, Num a) => (f (Sparse a) -> b) -> f a -> b

vars :: (Traversable f, Num a) => f a -> f (Sparse a)

d :: (Traversable f, Num a) => f b -> Sparse a -> f a

d' :: (Traversable f, Num a) => f a -> Sparse a -> (a, f a)

ds :: (Traversable f, Num a) => f b -> Sparse a -> Cofree f a

skeleton :: Traversable f => f a -> f Int

spartial :: Num a => [Int] -> Sparse a -> Maybe a

partial :: Num a => [Int] -> Sparse a -> a

vgrad :: Grad i o o' a => i -> o

vgrad' :: Grad i o o' a => i -> o'

vgrads :: Grads i o a => i -> o

class Num a => Grad i o o' a | i -> a o o', o -> a i o', o' -> a i o where

Methods

pack :: i -> [Sparse a] -> Sparse a

unpack :: ([a] -> [a]) -> o

unpack' :: ([a] -> (a, [a])) -> o'

class Num a => Grads i o a | i -> a o, o -> a i where

Methods

packs :: i -> [Sparse a] -> Sparse a

unpacks :: ([a] -> Cofree [] a) -> o

terms :: Monomial -> [(Integer, Monomial, Monomial)]

The value of the derivative of (f*g) of order mi is

sum [a * primal (partialS (indices b) f) * primal (partialS (indices c) g) | (a,b,c) <- terms mi ]

It is a bit more complicated in mul below, since we build the whole tree of derivatives and want to prune the tree with Zeros as much as possible. The number of terms in the sum for order mi as of differentiation has sum (map (+1) as) terms, so this is *much* more efficient than the naive recursive differentiation with 2^sum as terms. The coefficients a, which collect equivalent derivatives, are suitable products of binomial coefficients.

primal :: Num a => Sparse a -> a